+++ migrations/2025-09-07-185241_record-ingest-abort-error/up.sql
@@ -1,0 +1,2 @@
+alter table info.ingests
+    add column message text;