+++ mmoldb-db/src/schema/info_schema.rs
@@ -62,6 +62,7 @@ pub mod info {
             finished_at -> Nullable<Timestamp>,
             aborted_at -> Nullable<Timestamp>,
             start_next_ingest_at_page -> Nullable<Text>,
+            message -> Nullable<Text>,
         }
     }